fix: usar BIGINT para IDs de Telegram, auto-init de schema y variables de n8n para notificaciones
- chat_id, message_id y reply_to_message_id a BIGINT (los IDs de
  Telegram exceden el rango de INTEGER)
- montar db/schema.sql en /docker-entrypoint-initdb.d/ para que se
  auto-ejecute si se recrea el volumen de postgres desde cero
- exponer N8N_BLOCK_ENV_ACCESS_IN_NODE=false y TELEGRAM_USER_CHAT_ID
  al contenedor de n8n, requeridos por el nodo de notificación
- regex de TP/SL acepta dos puntos opcionales (TP: / SL:)
- parse_mode a "none" en el nodo de Telegram para evitar errores de
  parseo de entidades con datos dinámicos

Flujo completo verificado end-to-end: Telethon -> webhook -> parser
regex -> Postgres -> notificación Telegram.

+++ db/schema.sql
@@ -10,12 +10,12 @@ CREATE TABLE IF NOT EXISTS signals (
     id                  SERIAL PRIMARY KEY,
 
     -- Datos de captura (sección 3.1)
-    message_id          INTEGER NOT NULL,
-    chat_id             INTEGER NOT NULL,
+    message_id          BIGINT NOT NULL,
+    chat_id             BIGINT NOT NULL,
     sender               TEXT,
     raw_text            TEXT NOT NULL,
     signal_timestamp    TIMESTAMP NOT NULL,
-    reply_to_message_id INTEGER,
+    reply_to_message_id BIGINT,
 
     -- Datos interpretados (sección 4.2)
     instrument          TEXT NOT NULL,
@@ -71,7 +71,7 @@ CREATE TABLE IF NOT EXISTS signal_modifications (
     signal_id            INTEGER NOT NULL REFERENCES signals(id),
 
     -- Datos de captura del mensaje de seguimiento
-    message_id           INTEGER NOT NULL,
+    message_id           BIGINT NOT NULL,
     raw_text              TEXT NOT NULL,
 
     modification_type     TEXT NOT NULL CHECK (modification_type IN (
